+++ nodes/voice/hangup-call.html
@@ -1,6 +1,6 @@
 <script type="text/javascript">
   RED.nodes.registerType('hangup-call', {
-    category: 'twilio',
+    category: 'twilio voice',
     color: '#a6bbcf',
     defaults: {
       name: { value: '' },
