+++ nodes/voice/hangup.html
@@ -1,6 +1,6 @@
 <script type="text/javascript">
   RED.nodes.registerType('hangup', {
-    category: 'twilio',
+    category: 'twilio voice',
     color: '#a6bbcf',
     defaults: {
       name: { value: '' },
