+++ nodes/voice/reject.html
@@ -1,6 +1,6 @@
 <script type="text/javascript">
   RED.nodes.registerType('reject', {
-    category: 'twilio',
+    category: 'twilio voice',
     color: '#a6bbcf',
     defaults: {
       name: { value: '' },
