+++ settings.js
@@ -19,7 +19,7 @@ module.exports = {
   userDir: __dirname,
 
   // defines order of the categories in the editor palette
-  paletteCategories: ['subflows', 'twilio', 'input', 'output', 'function', 'storage', 'advanced'],
+  paletteCategories: ['subflows', 'twilio', 'twilio voice', 'input', 'output', 'function', 'storage', 'advanced'],
 
   // sets global context data
   functionGlobalContext: {
