+++ test/nodes/voice/play_spec.js
@@ -1,14 +1,13 @@
 var should = require('should');
 var helper = require('node-red-node-test-helper');
-var shared = require('./shared.js');
-var ttsNode = require('../../nodes/config/tts.js');
-var sayNode = require('./../../nodes/say.js');
-var res = require('./mocks.js').res;
+var shared = require('../shared.js');
+var playNode = require('../../../nodes/voice/play.js');
+var res = require('../mocks.js').res;
 var fs = require('fs');
 
 helper.init(require.resolve('node-red'));
 
-describe('say node', function() {
+describe('play node', function() {
   beforeEach(function(done) {
     helper.startServer(done);
   });
@@ -18,16 +17,12 @@ describe('say node', function() {
     helper.stopServer(done);
   });
 
-  shared.shouldLoadCorrectly(sayNode, 'say');
+  shared.shouldLoadCorrectly(playNode, 'play');
 
   it('should respond with proper XML', function(done) {
-    var flow = [
-      { id: 'n1', type: 'say', tts: 'n2', output: 'off' },
-      { id: 'n2', type: 'tts', text: 'hello', voice: 'alice', language: 'en-US' },
-    ];
-    var testNodes = [ttsNode, sayNode];
-    var xml = fs.readFileSync('test/resources/xml/say.xml', 'utf8');
-    helper.load(testNodes, flow, function() {
+    var flow = [{ id: 'n1', type: 'play', url: 'http://example.com/example.wav', output: 'off' }];
+    var xml = fs.readFileSync('test/resources/xml/play.xml', 'utf8');
+    helper.load(playNode, flow, function() {
       var n1 = helper.getNode('n1');
       n1.on('input', function(msg) {
         should(msg.res._res.responseBody).be.eql(xml);
