+++ test/nodes/voice/say_spec.js
@@ -1,13 +1,14 @@
 var should = require('should');
 var helper = require('node-red-node-test-helper');
-var shared = require('./shared.js');
-var playNode = require('./../../nodes/play.js');
-var res = require('./mocks.js').res;
+var shared = require('../shared.js');
+var ttsNode = require('../../../nodes/config/tts.js');
+var sayNode = require('../../../nodes/voice/say.js');
+var res = require('../mocks.js').res;
 var fs = require('fs');
 
 helper.init(require.resolve('node-red'));
 
-describe('play node', function() {
+describe('say node', function() {
   beforeEach(function(done) {
     helper.startServer(done);
   });
@@ -17,12 +18,16 @@ describe('play node', function() {
     helper.stopServer(done);
   });
 
-  shared.shouldLoadCorrectly(playNode, 'play');
+  shared.shouldLoadCorrectly(sayNode, 'say');
 
   it('should respond with proper XML', function(done) {
-    var flow = [{ id: 'n1', type: 'play', url: 'http://example.com/example.wav', output: 'off' }];
-    var xml = fs.readFileSync('test/resources/xml/play.xml', 'utf8');
-    helper.load(playNode, flow, function() {
+    var flow = [
+      { id: 'n1', type: 'say', tts: 'n2', output: 'off' },
+      { id: 'n2', type: 'tts', text: 'hello', voice: 'alice', language: 'en-US' },
+    ];
+    var testNodes = [ttsNode, sayNode];
+    var xml = fs.readFileSync('test/resources/xml/say.xml', 'utf8');
+    helper.load(testNodes, flow, function() {
       var n1 = helper.getNode('n1');
       n1.on('input', function(msg) {
         should(msg.res._res.responseBody).be.eql(xml);
